refactor(test): migrate runner and export helpers to three-tier architecture
Move DB-direct functions from api-test-helpers/runner.ts (3 seeders, 1
assertion) and api-test-helpers/export.ts (1 seeder, 1 assertion) to their
respective db-test-seeders/ and db-test-assertions/ files. Replace originals
with pure re-exports for backward compatibility.

Closes #9367
