+++ turbo/apps/web/src/__tests__/db-test-assertions/export.ts
@@ -1,0 +1,17 @@
+import { eq } from "drizzle-orm";
+import { initServices } from "../../lib/init-services";
+import { exportJobs } from "../../db/schema/export-job";
+
+/**
+ * Find an export job by ID.
+ * Returns the full row or null if not found.
+ */
+export async function findTestExportJobById(id: string) {
+  initServices();
+  const [row] = await globalThis.services.db
+    .select()
+    .from(exportJobs)
+    .where(eq(exportJobs.id, id))
+    .limit(1);
+  return row ?? null;
+}
